Added UI and backend for creating/editing variables, as well as for viewing and answering questions.
