+++ src/main/java/dstobns/ElGameAPI.java
@@ -1,5 +1,8 @@
 package dstobns;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import com.google.api.server.spi.config.Api;
 import com.google.api.server.spi.config.ApiMethod;
 import com.google.api.server.spi.config.ApiMethod.HttpMethod;
@@ -10,10 +13,11 @@ import com.google.appengine.api.users.User;
 import static dstobns.OfyService.ofy;
 
 import com.googlecode.objectify.Key;
+import com.googlecode.objectify.cmd.Query;
 
+import dstobns.domain.Answer;
 import dstobns.domain.BNVariable;
 import dstobns.domain.Profile;
-import dstobns.forms.NewVariableForm;
 
 import javax.inject.Named;
 
@@ -51,39 +55,90 @@ public class ElGameAPI {
 	  
   }
   
-  @ApiMethod(name = "storeVariable", path = "storeVariable", httpMethod = "post")
-  public BNVariable storeVariable(final User user, BNVariable variable) throws UnauthorizedException {
-	  
-	  if (user == null) {
-		  throw new UnauthorizedException("Authorization required");
-	  }
-	  
-	  // Save the entity in the datastore
-      ofy().save().entity(variable).now();
-	  
-	  return variable;
-	  
-  }
-  
-  // Declare this method as a method available externally through Endpoints
-  @ApiMethod(name = "saveVariable", path = "put-variable", httpMethod = HttpMethod.POST)
+  @ApiMethod(
+		  name = "saveVariable", 
+		  path = "put-variable", 
+		  httpMethod = HttpMethod.POST
+		  )
   // The request that invokes this method should provide data that
   // conforms to the fields defined in NewVariableForm
-  public BNVariable saveVariable(final User user, NewVariableForm newVariableForm)
-          throws UnauthorizedException {
+  public BNVariable saveVariable(final User user, 
+		  @Named("uniqueId") Long uniqueId, 
+		  @Named("name") String name, 
+		  @Named("label") String label, 
+		  @Named("states") String states
+		  )
+				  throws UnauthorizedException {
 
       // If the user is not logged in, throw an UnauthorizedException
       if (user == null) {
           throw new UnauthorizedException("Authorization required");
       }
       
-      BNVariable variable = new BNVariable(newVariableForm); 
+      BNVariable variable = new BNVariable(uniqueId, name, label, states); 
       
       // Save the entity in the datastore
       ofy().save().entity(variable).now();
 
       // Return the profile
       return variable;
+  }
+  
+  @ApiMethod(name = "retrieveVariables", path = "get-variables", httpMethod = HttpMethod.GET)
+  // Retrieve all variables 
+  public List<BNVariable> retrieveVariables(final User user)
+		  throws UnauthorizedException {
+
+      // If the user is not logged in, throw an UnauthorizedException
+      if (user == null) {
+          throw new UnauthorizedException("Authorization required");
+      }
+
+      return ofy().load().type(BNVariable.class).list();
+  }
+  
+  @ApiMethod(
+		  name = "retrieveAnswers", 
+		  path = "get-answers", 
+		  httpMethod = HttpMethod.GET
+		  )
+  // The request that invokes this method should provide data that
+  // conforms to the fields defined in NewVariableForm
+  public List<Answer> retrieveAnswers(final User user)
+				  throws UnauthorizedException {
+
+      // If the user is not logged in, throw an UnauthorizedException
+      if (user == null) {
+          throw new UnauthorizedException("Authorization required");
+      }
+      
+      List<BNVariable> variables = ofy().load().type(BNVariable.class).list(); 
+      
+      Answer answer; 
+      List<Answer> answers = new ArrayList<Answer>(); 
+      
+      for(BNVariable variableCause : variables) {
+    	  for(BNVariable variableEffect : variables) {
+    		  
+    		  answer = ofy().load().type(Answer.class)
+    				  .filter("userId", user.getUserId())
+    				  .filter("bnvariableCause", variableCause.getUniqueId())
+    				  .filter("bnvariableEffect", variableEffect.getUniqueId())
+    				  .first().now();
+    		  
+    		  if(answer == null) {
+    			  
+    			  answer = new Answer(user.getUserId(), variableCause.getUniqueId(), variableEffect.getUniqueId(), null); 
+    			  // Save the entity in the datastore 
+    		      ofy().save().entity(answer).now(); 
+    		  }
+    		  
+    		  answers.add(answer); 
+    		  
+    	  }
+      }
+      
+      return answers;
   }
 
   @ApiMethod(name = "getProfile", path = "getProfile")
