+++ src/main/webapp/index.html
@@ -2,8 +2,9 @@
 <html>
 <head>
     <title>The Elicitation Game!</title>
-    
+	
     <link type="text/css" rel="stylesheet" href="bower_components/bootstrap/dist/css/bootstrap.css">
+    <link rel="stylesheet" href="bower_components/angular-ui-select/dist/select.css">
     <link rel="shortcut icon" href="img/favicon.ico">
 </head>
 
@@ -31,8 +32,7 @@
 	        <li class="dropdown">
 	          <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-expanded="false">Admin <span class="caret"></span></a>
 	          <ul class="dropdown-menu" role="menu">
-	            <li><a href="#/new_variable">Create New Variable</a></li>
-	            <li><a href="#/edit_variables">Edit Variables</a></li>
+	            <li><a href="#/edit_variables">Edit/Create Variable</a></li>
 	            <li><a href="#">Something else here</a></li>
 	            <li class="divider"></li>
 	            <li><a href="#">Separated link</a></li>
@@ -68,21 +68,24 @@
 	</div>
 
 	<div class="container">
-	  	
+	
 	  	<script type="text/javascript" src="bower_components/angular/angular.js"></script>
 		<script type="text/javascript" src="bower_components/angular-route/angular-route.js"></script>
+		<script type="text/javascript" src="bower_components/angular-sanitize/angular-sanitize.js"></script>
+		<script type="text/javascript" src="bower_components/angular-ui-select/dist/select.js"></script>
 		<script type="text/javascript" src="bower_components/jquery/dist/jquery.js"></script>
 	    <script type="text/javascript" src="bower_components/bootstrap/dist/js/bootstrap.js"></script>
-	  	<script type="text/javascript" src="js/ui-bootstrap/ui-bootstrap-tpls-0.9.0.js"></script>
+	  	<script type="text/javascript" src="js/ui-bootstrap/ui-bootstrap-tpls-0.12.1.js"></script>
 	    <script type="text/javascript" src="js/base.js"></script>
 	    <script type="text/javascript" src="js/controllers.js"></script>
+	    <script type="text/javascript" src="js/questions.js"></script> 
 	  	
 	    <script type="text/javascript">
 	        function init() {
 	            elgame.init('//' + window.location.host + '/_ah/api');
 	        }
 	    </script>
-	    <script src="https://apis.google.com/js/client:platform.js?onload=init" async defer></script>
+	    <script src="https://apis.google.com/js/client:platform.js?onload=init"></script>
 	</div>
 	
 	<!-- Put the signInButton to invoke the gapi.signin.render to restore the credential if stored in cookie. -->
