+++ src/main/webapp/partials/edit_variables.html
@@ -1,4 +1,4 @@
-<div ng-controller="NewVariableCtrl">
+<div ng-controller="EditVariablesCtrl">
     <div class="row">
         <div class="col-lg-12">
             <div id="messages" class="alert alert-{{alertStatus}}" ng-show="messages">
@@ -12,13 +12,33 @@
 	<div class="row">
         <div class="col-md-8">
             <h3>Create a new variable</h3>
+  			
+  			<ui-select 	ng-model="bnvariableAsync.selected" 
+  						theme="bootstrap" 
+  						ng-disabled="disabled" 
+  						style="min-width: 300px;" 
+  						title="Select variable">
+	    		<ui-select-match placeholder="Select/search for a variable in the list or create a new variable...">{{$select.selected.name}}</ui-select-match>
+	   			<ui-select-choices repeat="bnvariable as bnvariable in bnvariablesAsync | propsFilter: {name: $select.search, id: $select.search}">
+			    	<div ng-bind-html="bnvariable.name | highlight: $select.search"></div>
+			    		<small>
+			        		label: {{bnvariable.label}}
+			        		id: <span ng-bind-html="''+bnvariable.uniqueId | highlight: $select.search"></span>
+			      		</small>
+	    		</ui-select-choices>
+  			</ui-select>
 
+			<hr>
+			
             <form name="NewVariableForm" novalidate role="form">
+            
+            	<input ng-show="false" name="uniqueId" type="text" ng-model="bnvariableAsync.selected.uniqueId" readonly/>
+            	
                 <div class="form-group">
                     <label for="name">Name <span class="required">*</span></label>
                     <span class="label label-danger"
                         ng-show="NewVariableForm.name.$error.required">Required!</span>
-                    <input id="name" type="text" name="name" ng-model="bnvariable.name" class="form-control"
+                    <input id="name" type="text" name="name" ng-model="bnvariableAsync.selected.name" class="form-control"
                            ng-required="true"/>
                 </div>
                 
@@ -26,7 +46,7 @@
                     <label for="label">Label <span class="required">*</span></label>
                     <span class="label label-danger"
                         ng-show="NewVariableForm.label.$error.required">Required!</span>
-                    <input id="label" type="text" name="label" ng-model="bnvariable.label" class="form-control"
+                    <input id="label" type="text" name="label" ng-model="bnvariableAsync.selected.label" class="form-control"
                            ng-required="true"/>
                 </div>
 
@@ -34,12 +54,12 @@
                     <label for="states">States <span class="required">*</span></label>
                     <span class="label label-danger"
                         ng-show="NewVariableForm.states.$error.required">Required!</span>
-                    <input id="states" type="text" name="states" ng-model="bnvariable.states" class="form-control"
+                    <input id="states" type="text" name="states" ng-model="bnvariableAsync.selected.states" class="form-control"
                            ng-required="true"/>
                 </div>
 
                 <button ng-click="saveVariable(NewVariableForm)" class="btn btn-primary"
-                        ng-disabled="!isValidVariable(NewVariableForm) || loading">Create
+                        ng-disabled="!isValidVariable(NewVariableForm) || loading">Save
                 </button>
             </form>
         </div>
