--- src/main/webapp/partials/new_variable.html
@@ -1,47 +1,0 @@
-<div ng-controller="NewVariableCtrl">
-    <div class="row">
-        <div class="col-lg-12">
-            <div id="messages" class="alert alert-{{alertStatus}}" ng-show="messages">
-                <span ng-bind="messages"></span>
-                <i class="dismiss-messages pull-right glyphicon glyphicon-remove" ng-click="messages = ''"
-                   ng-show="messages"></i>
-            </div>
-            <img class="spinner" src="/img/ajax-loader.gif" ng-show="loading"/>
-        </div>
-    </div>
-	<div class="row">
-        <div class="col-md-8">
-            <h3>Create a new variable</h3>
-
-            <form name="NewVariableForm" novalidate role="form">
-                <div class="form-group">
-                    <label for="name">Name <span class="required">*</span></label>
-                    <span class="label label-danger"
-                        ng-show="NewVariableForm.name.$error.required">Required!</span>
-                    <input id="name" type="text" name="name" ng-model="bnvariable.name" class="form-control"
-                           ng-required="true"/>
-                </div>
-                
-                <div class="form-group">
-                    <label for="label">Label <span class="required">*</span></label>
-                    <span class="label label-danger"
-                        ng-show="NewVariableForm.label.$error.required">Required!</span>
-                    <input id="label" type="text" name="label" ng-model="bnvariable.label" class="form-control"
-                           ng-required="true"/>
-                </div>
-
-                <div class="form-group">
-                    <label for="states">States <span class="required">*</span></label>
-                    <span class="label label-danger"
-                        ng-show="NewVariableForm.states.$error.required">Required!</span>
-                    <input id="states" type="text" name="states" ng-model="bnvariable.states" class="form-control"
-                           ng-required="true"/>
-                </div>
-
-                <button ng-click="saveVariable(NewVariableForm)" class="btn btn-primary"
-                        ng-disabled="!isValidVariable(NewVariableForm) || loading">Create
-                </button>
-            </form>
-        </div>
-    </div>
-</div>
